@@ -1,13 +1,10 @@
 ﻿using System.Collections.Generic;
 
-namespace MegaDTelegramRemoteControl.Infrastructure.Configurations
+namespace MegaDTelegramRemoteControl.Infrastructure.Configurations;
+
+public record TelegramConfig
 {
-    public record TelegramConfig
-    {
-        public string BotAccessToken { get; init; } = null!;
+    public string BotAccessToken { get; init; } = null!;
 
-        public HashSet<int> AllowedUsers { get; init; } = new();
-
-        public HashSet<int> DebugLogUsers { get; init; } = new();
-    }
+    public HashSet<int> AllowedUsers { get; init; } = new();
 }
