@@ -1,15 +1,14 @@
 ﻿using System;
 
-namespace MegaDTelegramRemoteControl.Models.Device
+namespace MegaDTelegramRemoteControl.Models.Device;
+
+[AttributeUsage(AttributeTargets.Field)]
+public class DeviceEventAttribute : Attribute
 {
-    [AttributeUsage(AttributeTargets.Field)]
-    public class DeviceEventAttribute : Attribute
+    public string Command { get; }
+
+    public DeviceEventAttribute(string command)
     {
-        public string Command { get; }
-        
-        public DeviceEventAttribute(string command)
-        {
-            Command = command;
-        }
+        Command = command;
     }
 }
