@@ -3,46 +3,45 @@ using MegaDTelegramRemoteControl.Infrastructure.Models;
 using System;
 using System.Text;
 
-namespace MegaDTelegramRemoteControl.Models.Device
+namespace MegaDTelegramRemoteControl.Models.Device;
+
+public class DevicePortEvent
 {
-    public class DevicePortEvent
+    private DevicePortInEvent? inPortData;
+    private DevicePortOutEvent? outPortData;
+
+    public DevicePort Port { get; init; } = null!;
+
+    public DateTime Date { get; init; }
+
+    public DevicePortInEvent? In
     {
-        private DevicePortInEvent? inPortData;
-        private DevicePortOutEvent? outPortData;
+        get => Port.Type == DevicePortType.IN
+            ? inPortData
+            : throw new OperationException($"'IN' port data is not available for port type {Port?.Type}");
+        set => inPortData = value;
+    }
 
-        public DevicePort Port { get; init; } = null!;
-        
-        public DateTime Date { get; init; }
+    public DevicePortOutEvent? Out
+    {
+        get => Port.Type == DevicePortType.OUT
+            ? outPortData
+            : throw new OperationException($"'Out' port data is not available for port type {Port?.Type}");
+        set => outPortData = value;
+    }
 
-        public DevicePortInEvent? In
-        {
-            get => Port.Type == DevicePortType.IN
-                ? inPortData
-                : throw new OperationException($"'IN' port data is not available for port type {Port?.Type}");
-            set => inPortData = value;
-        }
+    public override string ToString()
+    {
+        var str = new StringBuilder();
 
-        public DevicePortOutEvent? Out
-        {
-            get => Port.Type == DevicePortType.OUT
-                ? outPortData
-                : throw new OperationException($"'Out' port data is not available for port type {Port?.Type}");
-            set => outPortData = value;
-        }
+        str.Append($"port: {Port.Description ?? Port.Name} ({Port.Type})");
 
-        public override string ToString()
-        {
-            var str = new StringBuilder();
+        if (inPortData != null)
+            str.Append($", {inPortData}");
 
-            str.Append($"port: {Port.Name} ({Port.Type})");
+        if (outPortData != null)
+            str.Append($", {outPortData}");
 
-            if (inPortData != null)
-                str.Append($", {inPortData}");
-
-            if (outPortData != null)
-                str.Append($", {outPortData}");
-            
-            return str.ToString();
-        }
+        return str.ToString();
     }
 }
