@@ -1,12 +1,12 @@
-﻿using MegaDTelegramRemoteControl.Models.Device;
+﻿using MegaDTelegramRemoteControl.Models;
+using MegaDTelegramRemoteControl.Models.Device;
 using System.Collections.Generic;
 
-namespace MegaDTelegramRemoteControl.Services.Interfaces
+namespace MegaDTelegramRemoteControl.Services.Interfaces;
+
+public interface IDeviceCommandParser
 {
-    public interface IDeviceCommandParser
-    {
-        DeviceEvent ParseEvent(string deviceId, IReadOnlyCollection<(string key, string value)> eventData);
-        
-        DevicePortStatus ParseStatus(DevicePort port, string portStatus);
-    }
+    DeviceEvent ParseEvent(string deviceId, IReadOnlyCollection<NewEventData> eventData);
+
+    DevicePortStatus ParseStatus(DevicePort port, string portStatus);
 }
